finished day 06

+++ 2023/daily-template/src/part2.rs
@@ -1,6 +1,6 @@
 use color_eyre::Result;
 
-pub fn process(input: &str) -> Result<u32> {
+pub fn process(input: &str) -> Result<usize> {
     todo!("day xx - part 2");
     Ok(0)
 }
